zed: Move common flags to root
Attach flags -q, -lake, -configdir, and -use to the zed root command.
This enables use to specify these flags before the subcommand:
zed -lake $HOME/.lake ls

Closes #3843
