+++ cmd/zed/serve/command.go
@@ -38,7 +38,6 @@ Requests may be issued to this service via the "zed api" command.
 
 type Command struct {
 	*root.Command
-	cli.LakeFlags
 	conf    service.Config
 	logger  *zap.Logger
 	logconf logger.Config
@@ -65,7 +64,6 @@ func New(parent charm.Command, f *flag.FlagSet) (charm.Command, error) {
 	f.Var(&c.logconf.Mode, "log.filemode", "logger file write mode (values: append, truncate, rotate)")
 	f.StringVar(&c.portFile, "portfile", "", "write listen port to file")
 	f.StringVar(&c.rootContentFile, "rootcontentfile", "", "file to serve for GET /")
-	c.LakeFlags.SetFlagsWithDefaultLake(f, "")
 	return c, nil
 }
 
@@ -75,6 +73,9 @@ func (c *Command) Run(args []string) error {
 		return err
 	}
 	defer cleanup()
+	if !c.LakeSpecified {
+		c.Lake = ""
+	}
 	uri, err := c.URI()
 	if err != nil {
 		return err
